provider/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.10.4
itsdangerous==0.24
mysql==0.0.1
wsgiref==0.1.2
